Remove warning
Changing this to type int removes the warning that both sides are
different type (size_t vs. int). This method was recommended by our
book, but I will do as VS suggests.

+++ Season.cpp
@@ -14,7 +14,7 @@ Season::Season()
 {
 }
 
-Season::Season(int length) : seasonLength{length}
+Season::Season(size_t length) : seasonLength{length}
 {
 }
